app/requirements.txt
Pillow==5.3.0
pytesseract==0.2.5
Wand==0.4.5
numpy==1.15.4
flask==1.0.2
